Update init_assets.lua
I think i found the problem with the Roboland portal. It is not declared in assest. See if it is correct what i changed, I cannot test portal-related issues at the moment. If you can test it or check it and you think is fine maybe you should upload this to steam asap. We can just add the lines for robo instead of replace the chezz lines if you like it better that way